pre-dev: вынес общие типы, константы и утилиты в отдельные файлы
