+++ vue/src/components/parts/game/GameBoard.vue
@@ -50,31 +50,12 @@
 
 <script setup lang="ts">
 import { onMounted, ref } from "vue"
-
-interface Gem {
-  id: number
-  type: number
-  row: number
-  color: string
-  col: number
-  flag?: boolean
-}
-
-const AVAILABLE_SIZES = [4, 5, 6, 7, 8]
-const MAX = 8
-const GEM_COLORS = [
-  '#FF6B6B', // Красный
-  '#4ECDC4', // Бирюзовый
-  '#45B7D1', // Голубой
-  '#96CEB4', // Зеленый
-  '#FFEAA7', // Желтый
-  '#DDA0DD', // Сливовый
-  '#98D8C8', // Мятный
-  '#F7DC6F'  // Светло-желтый
-]
+import { GemType } from "@/types"
+import { findHorizontalMatches, findVerticalMatches } from "@/services/utils"
+import { GEM_COLORS, AVAILABLE_SIZES } from "@/services/constants"
 
 const gridSize = ref(8)
-const rows = ref<Gem[][]>([])
+const rows = ref<GemType[][]>([])
 const nextGemId = ref(1)
 
 const generateGemId = () => nextGemId.value++
@@ -91,7 +72,7 @@ const createGem = (row: number, col: number) => {
 }
 const setBoard = () => {
   const size = gridSize.value
-  const newBoard: Gem[][] = []
+  const newBoard: GemType[][] = []
 
   nextGemId.value = 1
 
@@ -107,15 +88,15 @@ const setBoard = () => {
   markMatches()
 }
 const markMatches = () => {
-  const matches: Gem[] = findMatches(rows.value)
+  const matches: GemType[] = findMatches(rows.value)
 
-  const newRows: Gem[][] = [ ...rows.value ]
-  matches.forEach((match: Gem) => {
+  const newRows: GemType[][] = [ ...rows.value ]
+  matches.forEach((match: GemType) => {
     newRows[match.row][match.col].flag = true
   })
 }
-const findMatches = (items: Gem[][]) => {
-  const matches: Gem[] = []
+const findMatches = (items: GemType[][]) => {
+  const matches: GemType[] = []
   const size = items.length
 
   const horizontalMatches = findHorizontalMatches(items, size)
@@ -123,62 +104,6 @@ const findMatches = (items: Gem[][]) => {
 
   const verticalMatches = findVerticalMatches(items, size)
   matches.push(...verticalMatches)
-
-  return matches
-}
-const findHorizontalMatches = (items: Gem[][], size: number) => {
-  const matches = []
-
-  for (let rowIndex = 0; rowIndex < size; rowIndex++) {
-    let count = 1
-
-    for (let colIndex = 1; colIndex < size; colIndex++) {
-      if (items[rowIndex][colIndex].type === items[rowIndex][colIndex - 1].type) {
-        count++
-      } else {
-        if (count >= 3) {
-          for(let countedIndex = colIndex - count; countedIndex < colIndex; countedIndex++) {
-            matches.push(items[rowIndex][countedIndex])
-          }
-        }
-        count = 1
-      }
-    }
-
-    if (count >= 3) {
-      for (let countedIndex = size - count; countedIndex < size; countedIndex++) {
-        matches.push(items[rowIndex][countedIndex])
-      }
-    }
-  }
-
-  return matches
-}
-const findVerticalMatches = (items: Gem[][], size: number) => {
-  const matches = []
-
-  for (let cIndex = 0; cIndex < size; cIndex++) {
-    let count = 1
-
-    for (let rIndex = 1; rIndex < size; rIndex++) {
-      if (items[rIndex][cIndex].type === items[rIndex - 1][cIndex].type) {
-        count++
-      } else {
-        if (count >= 3) {
-          for(let countedIndex = rIndex - count; countedIndex < rIndex; countedIndex++) {
-            matches.push(items[countedIndex][cIndex])
-          }
-        }
-        count = 1
-      }
-    }
-
-    if (count >= 3) {
-      for (let countedIndex = size - count; countedIndex < size; countedIndex++) {
-        matches.push(items[countedIndex][cIndex])
-      }
-    }
-  }
 
   return matches
 }
