+++ vue/src/services/constants.ts
@@ -1,0 +1,12 @@
+export const GEM_COLORS = [
+  '#FF6B6B', // Красный
+  '#4ECDC4', // Бирюзовый
+  '#45B7D1', // Голубой
+  '#96CEB4', // Зеленый
+  '#FFEAA7', // Желтый
+  '#DDA0DD', // Сливовый
+  '#98D8C8', // Мятный
+  '#F7DC6F'  // Светло-желтый
+]
+
+export const AVAILABLE_SIZES = [4, 5, 6, 7, 8]
